PythonAPI/requirements.txt
deltalake==0.11.0
fastapi==0.103.1
uvicorn==0.15.0
pydantic-settings==2.0.3